+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to
   flexibility for conversions
 - _breaking_ removes array as a strategy, use `composeStrategy` instead
 - _breaking_ `composeStrategy`'s FromJSON and ToJSON strategy types are removed
-	with `Strategy` replacing both of them
+  with `Strategy` replacing both of them
 - fixes #141
 - fixes #139
 - move `docs/` to `README.md` and delete folder